% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mc_add_library_update}
\alias{mc_add_library_update}
\title{Add an update to an existing library}
\usage{
mc_add_library_update(conn, library, updatename, js_code)
}
\description{
Add an update to an existing library
}
