% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mc_add_library_update_from}
\alias{mc_add_library_update_from}
\title{Add an update to a library from a specific address}
\usage{
mc_add_library_update_from(conn, from_address, library, updatename, js_code)
}
\description{
Add an update to a library from a specific address
}
