% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_cache.R
\name{mc_append_binary_cache}
\alias{mc_append_binary_cache}
\title{Append data to a binary cache item}
\usage{
mc_append_binary_cache(conn, identifier, data = "")
}
\arguments{
\item{conn}{Connection object.}

\item{identifier}{The cache item identifier (from mc_create_binary_cache).}

\item{data}{Data to append. Can be a hex string, or a list like list(text="...") or list(json=...).}
}
\value{
The resulting size of the cache item in bytes.
}
\description{
Append data to a binary cache item
}
