% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_transactions.R
\name{mc_append_raw_change}
\alias{mc_append_raw_change}
\title{Add a change output to a raw transaction}
\usage{
mc_append_raw_change(conn, tx_hex, address, native_fee = NULL)
}
\arguments{
\item{conn}{Connection object.}

\item{tx_hex}{Hex string of the raw transaction.}

\item{address}{Address to receive the change.}

\item{native_fee}{Optional native fee amount.}
}
\description{
Add a change output to a raw transaction
}
