% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_transactions.R
\name{mc_append_raw_transaction}
\alias{mc_append_raw_transaction}
\title{Add inputs and outputs to a raw transaction}
\usage{
mc_append_raw_transaction(conn, tx_hex, inputs = list(), outputs = list())
}
\description{
Add inputs and outputs to a raw transaction
}
