% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet.R
\name{mc_change_wallet_passphrase}
\alias{mc_change_wallet_passphrase}
\title{Change the wallet passphrase}
\usage{
mc_change_wallet_passphrase(conn, old_passphrase, new_passphrase)
}
\arguments{
\item{conn}{A connection object to the MultiChain node.}

\item{old_passphrase}{Character. The current password.}

\item{new_passphrase}{Character. The new password.}
}
\value{
Invisibly returns the RPC result (typically \code{NULL}) on success.
}
\description{
Changes the encryption password of the wallet.
}
\seealso{
Other wallet: 
\code{\link{mc_backup_wallet}()},
\code{\link{mc_combine_unspent}()},
\code{\link{mc_dump_privkey}()},
\code{\link{mc_dump_wallet}()},
\code{\link{mc_encrypt_wallet}()},
\code{\link{mc_get_wallet_info}()},
\code{\link{mc_import_privkey}()},
\code{\link{mc_import_wallet}()},
\code{\link{mc_list_lock_unspent}()},
\code{\link{mc_list_unspent}()},
\code{\link{mc_lock_unspent}()},
\code{\link{mc_lock_wallet}()},
\code{\link{mc_unlock_wallet}()}
}
\concept{wallet}
