% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{mc_connect}
\alias{mc_connect}
\title{Create a MultiChain connection object}
\usage{
mc_connect(host = "127.0.0.1", port = NULL, user = NULL, password = NULL)
}
\arguments{
\item{host}{IP address or a config list from mc_get_config.}

\item{port}{RPC port number.}

\item{user}{RPC username.}

\item{password}{RPC password.}
}
\value{
A \code{multichain_conn} object.
}
\description{
Create a MultiChain connection object
}
