% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addresses.R
\name{mc_create_keypairs}
\alias{mc_create_keypairs}
\title{Create new key pairs}
\usage{
mc_create_keypairs(conn, count = 1)
}
\arguments{
\item{conn}{Connection object.}

\item{count}{Number of key pairs to generate (default 1).}
}
\value{
A data frame with columns: address, pubkey, and privkey.
}
\description{
Generates one or more public/private key pairs. These are NOT stored
in the node's wallet.
}
