% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addresses.R
\name{mc_create_multisig}
\alias{mc_create_multisig}
\title{Create multi-signature address (external)}
\usage{
mc_create_multisig(conn, n_required, keys)
}
\arguments{
\item{conn}{Connection object.}

\item{n_required}{Number of required signatures.}

\item{keys}{A character vector of public keys or addresses.}
}
\value{
A list containing the 'address' and 'redeemScript'.
}
\description{
Creates a P2SH multi-signature address without adding it to the wallet.
}
