% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{mc_create_stream_from}
\alias{mc_create_stream_from}
\title{Create a new stream from a specific address}
\usage{
mc_create_stream_from(
  conn,
  from_address,
  name,
  open = TRUE,
  custom_fields = NULL
)
}
\arguments{
\item{conn}{Connection object.}

\item{from_address}{Address used to create the stream.}

\item{name}{Stream name.}

\item{open}{Boolean or list of parameters.}

\item{custom_fields}{Optional list of custom fields.}
}
\description{
Create a new stream from a specific address
}
