% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mc_create_upgrade}
\alias{mc_create_upgrade}
\title{Create a blockchain upgrade}
\usage{
mc_create_upgrade(conn, name, params)
}
\arguments{
\item{conn}{A connection object created by \code{\link{mc_connect}}.}

\item{name}{Character string. Name of the upgrade (must be unique).}

\item{params}{List of parameters to upgrade, e.g.,
\code{list("target-block-time" = 20, "max-block-size" = 10000000)}.}
}
\value{
A list containing the result of the RPC call (usually transaction ID).
}
\description{
Creates a new upgrade proposal to change blockchain parameters (e.g., target
block time, maximum block size). Upgrades require admin approval.
}
\examples{
\dontrun{
mc_create_upgrade(conn, "speedup", list("target-block-time" = 20))
}

}
\seealso{
\code{\link{mc_approve_from}} to approve an upgrade.

Other filters: 
\code{\link{mc_approve_from}()},
\code{\link{mc_create_stream_filter}()},
\code{\link{mc_create_tx_filter}()},
\code{\link{mc_get_filter_code}()},
\code{\link{mc_list_stream_filters}()},
\code{\link{mc_list_tx_filters}()},
\code{\link{mc_list_upgrades}()},
\code{\link{mc_run_stream_filter}()},
\code{\link{mc_run_tx_filter}()},
\code{\link{mc_test_stream_filter}()},
\code{\link{mc_test_tx_filter}()}
}
\concept{filters}
