% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mc_create_variable_from}
\alias{mc_create_variable_from}
\title{Create a variable from a specific address}
\usage{
mc_create_variable_from(conn, from_address, name, open = TRUE, value = NULL)
}
\description{
Create a variable from a specific address
}
