% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomic_exchange.R
\name{mc_decode_raw_exchange}
\alias{mc_decode_raw_exchange}
\title{Decode a raw exchange transaction}
\usage{
mc_decode_raw_exchange(conn, tx_hex, verbose = FALSE)
}
\arguments{
\item{conn}{Connection object.}

\item{tx_hex}{Hex string of the exchange transaction.}

\item{verbose}{If TRUE, lists all individual stages in the exchange.}
}
\description{
Decode a raw exchange transaction
}
