% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomic_exchange.R
\name{mc_disable_raw_transaction}
\alias{mc_disable_raw_transaction}
\title{Disable an offer of exchange}
\usage{
mc_disable_raw_transaction(conn, tx_hex)
}
\arguments{
\item{conn}{Connection object.}

\item{tx_hex}{Hex string of the exchange transaction.}
}
\value{
Transaction ID of the disabling transaction.
}
\description{
Disable an offer of exchange
}
