% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet.R
\name{mc_dump_privkey}
\alias{mc_dump_privkey}
\title{Dump a private key for an address}
\usage{
mc_dump_privkey(conn, address)
}
\arguments{
\item{conn}{A connection object to the MultiChain node.}

\item{address}{Character. The wallet address for which to retrieve the private key.}
}
\value{
Character. The private key in Wallet Import Format (WIF).
}
\description{
Dump a private key for an address
}
\seealso{
Other MultiChain Wallet Management: 
\code{\link{mc_backup_wallet}()},
\code{\link{mc_change_wallet_passphrase}()},
\code{\link{mc_dump_wallet}()},
\code{\link{mc_encrypt_wallet}()},
\code{\link{mc_get_wallet_info}()},
\code{\link{mc_import_privkey}()},
\code{\link{mc_import_wallet}()},
\code{\link{mc_lock_wallet}()},
\code{\link{mc_unlock_wallet}()}
}
\concept{MultiChain Wallet Management}
