% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{mc_get_added_node_info}
\alias{mc_get_added_node_info}
\title{Get information about manually added nodes}
\usage{
mc_get_added_node_info(conn, verbose = FALSE, node = NULL)
}
\arguments{
\item{conn}{Connection object.}

\item{verbose}{If TRUE, returns detailed information.}

\item{node}{Optional specific node address.}
}
\value{
A list or data frame of added nodes.
}
\description{
Get information about manually added nodes
}
