% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{mc_get_address_transaction}
\alias{mc_get_address_transaction}
\title{Get details of a transaction related to an address}
\usage{
mc_get_address_transaction(conn, address, txid, verbose = FALSE)
}
\arguments{
\item{conn}{Connection object.}

\item{address}{Wallet address.}

\item{txid}{Transaction ID.}

\item{verbose}{If TRUE, provides details of inputs and outputs.}
}
\description{
Get details of a transaction related to an address
}
