% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addresses.R
\name{mc_get_addresses}
\alias{mc_get_addresses}
\title{Get node wallet addresses}
\usage{
mc_get_addresses(conn, verbose = FALSE)
}
\arguments{
\item{conn}{Connection object.}

\item{verbose}{If TRUE, returns detailed information about addresses.}
}
\description{
Returns all addresses owned by the current node.
}
