% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{mc_get_asset_info}
\alias{mc_get_asset_info}
\title{Get information about a specific asset}
\usage{
mc_get_asset_info(conn, asset, verbose = FALSE)
}
\arguments{
\item{conn}{Connection object.}

\item{asset}{Asset name, ref or issuance txid.}

\item{verbose}{If TRUE, returns details about individual issuances.}
}
\description{
Get information about a specific asset
}
