% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockchain.R
\name{mc_get_block_hash}
\alias{mc_get_block_hash}
\title{Get block hash by height}
\usage{
mc_get_block_hash(conn, height)
}
\arguments{
\item{conn}{A connection object created by \code{\link{mc_connect}}.}

\item{height}{Integer. The block height (0 for genesis block).}
}
\value{
A character string containing the block hash.
}
\description{
Returns the hash of a block at a given height.
}
\examples{
\dontrun{
hash <- mc_get_block_hash(conn, 0)  # genesis block hash
}

}
\seealso{
\code{\link{mc_get_block}} to retrieve block details.

Other blockchain information: 
\code{\link{mc_get_block}()},
\code{\link{mc_get_blockchain_info}()},
\code{\link{mc_get_chain_totals}()},
\code{\link{mc_get_last_block_info}()},
\code{\link{mc_list_blocks}()},
\code{\link{mc_list_miners}()}
}
\concept{blockchain information}
