% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_runtime.R
\name{mc_get_chunk_queue_info}
\alias{mc_get_chunk_queue_info}
\title{Get information about off-chain chunk queue}
\usage{
mc_get_chunk_queue_info(conn)
}
\arguments{
\item{conn}{Connection object.}
}
\value{
A list containing chunk counts and byte sizes.
}
\description{
Get information about off-chain chunk queue
}
