% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_runtime.R
\name{mc_get_chunk_queue_totals}
\alias{mc_get_chunk_queue_totals}
\title{Get cumulative statistics on off-chain chunk requests}
\usage{
mc_get_chunk_queue_totals(conn)
}
\arguments{
\item{conn}{Connection object.}
}
\value{
A list of statistics (delivered, undelivered, timeouts, etc.).
}
\description{
Get cumulative statistics on off-chain chunk requests
}
