% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mc_get_filter_code}
\alias{mc_get_filter_code}
\title{Get JavaScript code of a filter}
\usage{
mc_get_filter_code(conn, filter)
}
\arguments{
\item{conn}{Connection object.}

\item{filter}{Filter.}
}
\description{
Get JavaScript code of a filter
}
