% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mc_get_library_code}
\alias{mc_get_library_code}
\title{Get JavaScript code for a library}
\usage{
mc_get_library_code(conn, library, updatename = NULL)
}
\arguments{
\item{updatename}{Optional. If omitted, returns active code. Use "" for initial code.}
}
\description{
Get JavaScript code for a library
}
