% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{mc_get_multi_balances}
\alias{mc_get_multi_balances}
\title{Get balances for multiple addresses and assets}
\usage{
mc_get_multi_balances(
  conn,
  addresses = "*",
  assets = "*",
  minconf = 1,
  include_watch_only = FALSE,
  include_locked = FALSE
)
}
\arguments{
\item{conn}{A connection object to the MultiChain node.}

\item{addresses}{A vector of addresses, or \code{"*"} for all addresses in the wallet (default \code{"*"}).}

\item{assets}{A vector of asset names/refs, or \code{"*"} for all assets (default \code{"*"}).}

\item{minconf}{Integer. Minimum confirmations (default \code{1}).}

\item{include_watch_only}{Logical. Include watch-only addresses (default \code{FALSE}).}

\item{include_locked}{Logical. Include locked unspent outputs (default \code{FALSE}).}
}
\value{
A list or data frame of balances, indexed by address and asset.
}
\description{
Returns a breakdown of balances across a set of addresses and/or assets.
}
\seealso{
Other transactions: 
\code{\link{mc_get_address_balances}()},
\code{\link{mc_get_address_transaction}()},
\code{\link{mc_get_token_balances}()},
\code{\link{mc_get_total_balances}()},
\code{\link{mc_get_tx_out_data}()},
\code{\link{mc_get_wallet_transaction}()},
\code{\link{mc_list_address_transactions}()},
\code{\link{mc_list_wallet_transactions}()},
\code{\link{mc_send}()},
\code{\link{mc_send_asset}()},
\code{\link{mc_send_asset_from}()},
\code{\link{mc_send_from}()},
\code{\link{mc_send_with_data}()},
\code{\link{mc_send_with_data_from}()}
}
\concept{transactions}
