% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addresses.R
\name{mc_get_new_address}
\alias{mc_get_new_address}
\title{Create new wallet address}
\usage{
mc_get_new_address(conn)
}
\arguments{
\item{conn}{Connection object.}
}
\description{
Create new wallet address
}
