% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{mc_get_peer_info}
\alias{mc_get_peer_info}
\title{Get information about connected peers}
\usage{
mc_get_peer_info(conn)
}
\arguments{
\item{conn}{Connection object.}
}
\value{
A data frame with peer details (latency, version, IP, etc.).
}
\description{
Get information about connected peers
}
