% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{mc_get_stream_publisher_summary}
\alias{mc_get_stream_publisher_summary}
\title{Get summary of JSON objects published by a specific address}
\usage{
mc_get_stream_publisher_summary(
  conn,
  stream,
  address,
  mode = "jsonobjectmerge"
)
}
\description{
Get summary of JSON objects published by a specific address
}
