% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{mc_get_token_balances}
\alias{mc_get_token_balances}
\title{Get balances for non-fungible tokens (NFTs)}
\usage{
mc_get_token_balances(
  conn,
  addresses = "*",
  assets = "*",
  minconf = 1,
  include_watch_only = FALSE,
  include_locked = FALSE
)
}
\arguments{
\item{conn}{A connection object to the MultiChain node.}

\item{addresses}{A vector of addresses, or \code{"*"} (default \code{"*"}).}

\item{assets}{A vector of parent asset names/refs, or \code{"*"} (default \code{"*"}).}

\item{minconf}{Integer. Minimum confirmations (default \code{1}).}

\item{include_watch_only}{Logical. Include watch-only addresses (default \code{FALSE}).}

\item{include_locked}{Logical. Include locked outputs (default \code{FALSE}).}
}
\value{
A data frame containing token-level balance details.
}
\description{
Specifically retrieves balances for tokens (sub-assets or individual units)
associated with a parent asset.
}
\seealso{
Other MultiChain Balances: 
\code{\link{mc_get_address_balances}()},
\code{\link{mc_get_multi_balances}()},
\code{\link{mc_get_total_balances}()}
}
\concept{MultiChain Balances}
