% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{mc_get_token_info}
\alias{mc_get_token_info}
\title{Get information about a specific token (MultiChain 2.2.1+)}
\usage{
mc_get_token_info(conn, asset, token, verbose = FALSE)
}
\arguments{
\item{conn}{Connection object.}

\item{asset}{Parent asset name, ref or issuance txid.}

\item{token}{Token name or index.}

\item{verbose}{If TRUE, returns detailed information.}
}
\description{
Get information about a specific token (MultiChain 2.2.1+)
}
