% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{mc_get_total_balances}
\alias{mc_get_total_balances}
\title{Get total wallet balances}
\usage{
mc_get_total_balances(
  conn,
  minconf = 1,
  include_watch_only = FALSE,
  include_locked = FALSE
)
}
\arguments{
\item{conn}{A connection object to the MultiChain node.}

\item{minconf}{Integer. Minimum confirmations (default \code{1}).}

\item{include_watch_only}{Logical. Include watch-only addresses (default \code{FALSE}).}

\item{include_locked}{Logical. Include locked outputs (default \code{FALSE}).}
}
\value{
A data frame summarizing total balances for each asset.
}
\description{
Returns the total balance of all assets across all addresses in the wallet.
}
\seealso{
Other transactions: 
\code{\link{mc_get_address_balances}()},
\code{\link{mc_get_address_transaction}()},
\code{\link{mc_get_multi_balances}()},
\code{\link{mc_get_token_balances}()},
\code{\link{mc_get_tx_out_data}()},
\code{\link{mc_get_wallet_transaction}()},
\code{\link{mc_list_address_transactions}()},
\code{\link{mc_list_wallet_transactions}()},
\code{\link{mc_send}()},
\code{\link{mc_send_asset}()},
\code{\link{mc_send_asset_from}()},
\code{\link{mc_send_from}()},
\code{\link{mc_send_with_data}()},
\code{\link{mc_send_with_data_from}()}
}
\concept{transactions}
