% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{mc_get_tx_out_data}
\alias{mc_get_tx_out_data}
\title{Retrieve full hex data from a transaction output}
\usage{
mc_get_tx_out_data(conn, txid, vout, count_bytes = NULL, start_byte = 0)
}
\description{
Useful for retrieving large stream data that was truncated.
}
