% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet.R
\name{mc_get_wallet_info}
\alias{mc_get_wallet_info}
\title{Get general wallet information}
\usage{
mc_get_wallet_info(conn)
}
\arguments{
\item{conn}{A connection object to the MultiChain node.}
}
\value{
A list of wallet information.
}
\description{
Returns metadata about the wallet, such as version, balance, and encryption status.
}
\seealso{
Other wallet: 
\code{\link{mc_backup_wallet}()},
\code{\link{mc_change_wallet_passphrase}()},
\code{\link{mc_combine_unspent}()},
\code{\link{mc_dump_privkey}()},
\code{\link{mc_dump_wallet}()},
\code{\link{mc_encrypt_wallet}()},
\code{\link{mc_import_privkey}()},
\code{\link{mc_import_wallet}()},
\code{\link{mc_list_lock_unspent}()},
\code{\link{mc_list_unspent}()},
\code{\link{mc_lock_unspent}()},
\code{\link{mc_lock_wallet}()},
\code{\link{mc_unlock_wallet}()}
}
\concept{wallet}
