% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{mc_get_wallet_transaction}
\alias{mc_get_wallet_transaction}
\title{Get details of a wallet transaction}
\usage{
mc_get_wallet_transaction(
  conn,
  txid,
  include_watch_only = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{conn}{Connection object.}

\item{txid}{Transaction ID.}

\item{include_watch_only}{Include watch-only addresses (default FALSE).}

\item{verbose}{If TRUE, provides details of inputs and outputs.}
}
\description{
Get details of a wallet transaction
}
