% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet.R
\name{mc_import_privkey}
\alias{mc_import_privkey}
\title{Import one or more private keys}
\usage{
mc_import_privkey(conn, privkeys, label = "", rescan = TRUE)
}
\arguments{
\item{conn}{Connection object.}

\item{privkeys}{A single private key string or a vector of strings.}

\item{label}{Optional label for the addresses.}

\item{rescan}{If TRUE (default), rescans the whole chain. Can be an integer block height.}
}
\description{
Import one or more private keys
}
