% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet.R
\name{mc_import_wallet}
\alias{mc_import_wallet}
\title{Import keys from a wallet dump file}
\usage{
mc_import_wallet(conn, filename, rescan = 0)
}
\arguments{
\item{conn}{Connection object.}

\item{filename}{Path to the dump file.}

\item{rescan}{Block number to start rescanning from (default 0).}
}
\description{
Import keys from a wallet dump file
}
