% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{mc_issue_from}
\alias{mc_issue_from}
\title{Issue new asset from specific address}
\usage{
mc_issue_from(
  conn,
  from_address,
  to_address,
  name,
  quantity,
  units = 1,
  native_amount = NULL,
  custom_fields = NULL
)
}
\arguments{
\item{conn}{Connection object.}

\item{from_address}{Address used to issue the asset.}

\item{to_address}{Recipient address.}

\item{name}{Asset name or params list.}

\item{quantity}{Amount to issue.}

\item{units}{Smallest divisible unit.}

\item{native_amount}{Optional native currency amount.}

\item{custom_fields}{Optional list of custom fields.}
}
\description{
Issue new asset from specific address
}
