% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{mc_list_asset_issues}
\alias{mc_list_asset_issues}
\title{List issuance events for an asset}
\usage{
mc_list_asset_issues(conn, asset, verbose = FALSE, count = NULL, start = NULL)
}
\arguments{
\item{conn}{Connection object.}

\item{asset}{Asset name, ref or txid.}

\item{verbose}{If TRUE, includes issuers and custom fields.}

\item{count}{Number of issuances to return.}

\item{start}{Offset (negative for most recent).}
}
\description{
List issuance events for an asset
}
