% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockchain.R
\name{mc_list_blocks}
\alias{mc_list_blocks}
\title{List information about specific blocks}
\usage{
mc_list_blocks(conn, blocks, verbose = FALSE)
}
\arguments{
\item{blocks}{Range (e.g. "100-200"), hash, height, or -n for most recent.}

\item{verbose}{If TRUE, returns detailed information.}
}
\description{
List information about specific blocks
}
