% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet.R
\name{mc_list_lock_unspent}
\alias{mc_list_lock_unspent}
\title{List locked unspent outputs}
\usage{
mc_list_lock_unspent(conn)
}
\arguments{
\item{conn}{Connection object.}
}
\value{
A data frame of locked outputs (txid and vout).
}
\description{
List locked unspent outputs
}
