% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{mc_list_stored_nodes}
\alias{mc_list_stored_nodes}
\title{List known peer node addresses (MultiChain 2.3+)}
\usage{
mc_list_stored_nodes(conn, include_old_ignores = FALSE)
}
\arguments{
\item{conn}{Connection object.}

\item{include_old_ignores}{Include old ignored nodes (default FALSE).}
}
\value{
A data frame of stored nodes.
}
\description{
List known peer node addresses (MultiChain 2.3+)
}
