% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mc_list_stream_filters}
\alias{mc_list_stream_filters}
\title{List stream filters}
\usage{
mc_list_stream_filters(conn, filters = "*", verbose = FALSE)
}
\arguments{
\item{conn}{Connection object.}

\item{filters}{Filters.}
}
\description{
List stream filters
}
