% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{mc_list_stream_items}
\alias{mc_list_stream_items}
\title{List items in a stream (Enhanced version)}
\usage{
mc_list_stream_items(
  conn,
  stream,
  verbose = FALSE,
  count = 10,
  start = NULL,
  local_ordering = FALSE
)
}
\description{
List items in a stream (Enhanced version)
}
