% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{mc_list_stream_key_items}
\alias{mc_list_stream_key_items}
\title{List items with a specific key}
\usage{
mc_list_stream_key_items(
  conn,
  stream,
  key,
  verbose = FALSE,
  count = 10,
  start = NULL,
  local_ordering = FALSE
)
}
\description{
List items with a specific key
}
