% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{mc_list_stream_publishers}
\alias{mc_list_stream_publishers}
\title{List publishers who have written to a stream}
\usage{
mc_list_stream_publishers(
  conn,
  stream,
  addresses = "*",
  verbose = FALSE,
  count = NULL,
  start = NULL,
  local_ordering = FALSE
)
}
\description{
List publishers who have written to a stream
}
