% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{mc_list_streams}
\alias{mc_list_streams}
\title{List streams on the blockchain}
\usage{
mc_list_streams(
  conn,
  streams = "*",
  verbose = FALSE,
  count = NULL,
  start = NULL
)
}
\arguments{
\item{conn}{Connection object.}

\item{streams}{Stream filter (name, ref, txid, vector or "*").}

\item{verbose}{If TRUE, returns detailed information.}

\item{count}{Number of streams to return.}

\item{start}{Offset.}
}
\description{
List streams on the blockchain
}
