% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mc_list_upgrades}
\alias{mc_list_upgrades}
\title{List upgrades}
\usage{
mc_list_upgrades(conn, upgrades = "*")
}
\arguments{
\item{conn}{A connection object created by \code{\link{mc_connect}}.}

\item{upgrades}{Character vector of upgrade names/IDs, or \code{"*"} (default)
for all upgrades.}
}
\value{
A data frame (via \code{rpc_res_to_df}) with upgrade information.
}
\description{
Returns a list of upgrade proposals on the blockchain.
}
\examples{
\dontrun{
upgrades <- mc_list_upgrades(conn)
}

}
\seealso{
Other upgrades: 
\code{\link{mc_create_upgrade}()}
}
\concept{upgrades}
