% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{mc_list_wallet_transactions}
\alias{mc_list_wallet_transactions}
\title{List transactions in the wallet}
\usage{
mc_list_wallet_transactions(
  conn,
  count = 10,
  skip = 0,
  include_watch_only = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{conn}{Connection object.}

\item{count}{Number of transactions (default 10).}

\item{skip}{Number of transactions to skip (default 0).}

\item{include_watch_only}{Include watch-only addresses (default FALSE).}

\item{verbose}{If TRUE, provides details of inputs and outputs.}
}
\description{
List transactions in the wallet
}
