% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet.R
\name{mc_lock_unspent}
\alias{mc_lock_unspent}
\title{Lock or unlock unspent outputs}
\usage{
mc_lock_unspent(conn, unlock, outputs = NULL)
}
\arguments{
\item{conn}{A connection object to the MultiChain node.}

\item{unlock}{Logical. If \code{TRUE}, unlocks the outputs; if \code{FALSE}, locks them.}

\item{outputs}{Optional list of outputs to (un)lock. Expected format:
\code{list(list(txid="id", vout=0), ...)}. If empty and \code{unlock=TRUE},
all outputs are unlocked.}
}
\value{
Logical \code{TRUE} on success.
}
\description{
Prevents specific UTXOs from being spent in automated transactions, or
releases them if they were previously locked.
}
\seealso{
Other MultiChain UTXO Management: 
\code{\link{mc_combine_unspent}()},
\code{\link{mc_list_lock_unspent}()},
\code{\link{mc_list_unspent}()}
}
\concept{MultiChain UTXO Management}
