% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet.R
\name{mc_lock_wallet}
\alias{mc_lock_wallet}
\title{Immediately lock the wallet}
\usage{
mc_lock_wallet(conn)
}
\arguments{
\item{conn}{A connection object to the MultiChain node.}
}
\value{
Invisibly returns the RPC result (typically \code{NULL}) on success.
}
\description{
Removes the wallet encryption key from memory, requiring a passphrase for
further private key operations.
}
\seealso{
Other wallet: 
\code{\link{mc_backup_wallet}()},
\code{\link{mc_change_wallet_passphrase}()},
\code{\link{mc_combine_unspent}()},
\code{\link{mc_dump_privkey}()},
\code{\link{mc_dump_wallet}()},
\code{\link{mc_encrypt_wallet}()},
\code{\link{mc_get_wallet_info}()},
\code{\link{mc_import_privkey}()},
\code{\link{mc_import_wallet}()},
\code{\link{mc_list_lock_unspent}()},
\code{\link{mc_list_unspent}()},
\code{\link{mc_lock_unspent}()},
\code{\link{mc_unlock_wallet}()}
}
\concept{wallet}
