% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_management.R
\name{mc_node_init}
\alias{mc_node_init}
\title{Initialize new MultiChain blockchain}
\usage{
mc_node_init(chain_name)
}
\arguments{
\item{chain_name}{Name of the blockchain to create.}
}
\description{
Initialize new MultiChain blockchain
}
