% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_local.R
\name{mc_node_start}
\alias{mc_node_start}
\title{Start a MultiChain node}
\usage{
mc_node_start(chain_name, datadir = NULL)
}
\arguments{
\item{chain_name}{Name of the blockchain to start.}

\item{datadir}{Optional custom data directory.}
}
\description{
Start a MultiChain node
}
