% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomic_exchange.R
\name{mc_prepare_lock_unspent_from}
\alias{mc_prepare_lock_unspent_from}
\title{Prepare an unspent output from a specific address}
\usage{
mc_prepare_lock_unspent_from(conn, from_address, amounts, lock = TRUE)
}
\arguments{
\item{conn}{Connection object.}

\item{from_address}{Sender address.}

\item{amounts}{List of assets or numeric amount.}

\item{lock}{If TRUE (default), the output is locked.}
}
\description{
Prepare an unspent output from a specific address
}
