% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{mc_publish}
\alias{mc_publish}
\title{Publish an item to a stream}
\usage{
mc_publish(conn, stream, keys, data, options = NULL)
}
\arguments{
\item{conn}{Connection object.}

\item{stream}{Stream name, ref or txid.}

\item{keys}{A single key (string) or a vector of keys.}

\item{data}{Data to publish. Can be a hex string, or a list like list(text = "...") or list(json = ...).}

\item{options}{Optional. Use "offchain" to publish as an off-chain item.}
}
\description{
Publish an item to a stream
}
