% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{mc_publish_from}
\alias{mc_publish_from}
\title{Publish an item to a stream from a specific address}
\usage{
mc_publish_from(conn, from_address, stream, keys, data, options = NULL)
}
\arguments{
\item{conn}{A connection object created by \code{\link{mc_connect}}.}

\item{from_address}{Character string. Address that will publish the item.}

\item{stream}{Character string. Stream name, reference, or txid.}

\item{keys}{A single key or vector of keys.}

\item{data}{Data to publish (string or list).}

\item{options}{Optional \code{"offchain"} string.}
}
\value{
A character string containing the transaction ID.
}
\description{
Similar to \code{\link{mc_publish}}, but specifies the publishing address.
}
\examples{
\dontrun{
mc_publish_from(conn, "1A...", "mystream", "key", list(text = "data"))
}

}
\seealso{
\code{\link{mc_publish}}

Other streams: 
\code{\link{mc_create_stream}()},
\code{\link{mc_create_stream_from}()},
\code{\link{mc_get_stream_info}()},
\code{\link{mc_get_stream_item}()},
\code{\link{mc_get_stream_key_summary}()},
\code{\link{mc_get_stream_publisher_summary}()},
\code{\link{mc_list_stream_block_items}()},
\code{\link{mc_list_stream_items}()},
\code{\link{mc_list_stream_key_items}()},
\code{\link{mc_list_stream_keys}()},
\code{\link{mc_list_stream_publisher_items}()},
\code{\link{mc_list_stream_publishers}()},
\code{\link{mc_list_stream_query_items}()},
\code{\link{mc_list_stream_tx_items}()},
\code{\link{mc_list_streams}()},
\code{\link{mc_publish}()},
\code{\link{mc_publish_multi}()},
\code{\link{mc_publish_multi_from}()}
}
\concept{streams}
