% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{mc_publish_multi_from}
\alias{mc_publish_multi_from}
\title{Publish multiple items from a specific address}
\usage{
mc_publish_multi_from(conn, from_address, stream, items, options = NULL)
}
\arguments{
\item{conn}{A connection object created by \code{\link{mc_connect}}.}

\item{from_address}{Character string. Address that will publish the items.}

\item{stream}{Character string. Default stream.}

\item{items}{List of item objects.}

\item{options}{Optional default \code{"offchain"} string.}
}
\value{
A character string containing the transaction ID.
}
\description{
Publishes multiple items in one transaction from a specified address.
}
\examples{
\dontrun{
items <- list(list(key = "k1", data = "value1"))
txid <- mc_publish_multi_from(conn, "1A...", "mystream", items)
}

}
\seealso{
\code{\link{mc_publish_multi}}

Other stream items: 
\code{\link{mc_get_stream_item}()},
\code{\link{mc_list_stream_block_items}()},
\code{\link{mc_list_stream_items}()},
\code{\link{mc_list_stream_key_items}()},
\code{\link{mc_list_stream_publisher_items}()},
\code{\link{mc_list_stream_tx_items}()},
\code{\link{mc_publish}()},
\code{\link{mc_publish_from}()},
\code{\link{mc_publish_multi}()}
}
\concept{stream items}
