% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{mc_revoke}
\alias{mc_revoke}
\title{Revoke permissions from an address}
\usage{
mc_revoke(conn, address, permissions)
}
\arguments{
\item{conn}{A multichain_conn object.}

\item{address}{Target wallet address.}

\item{permissions}{Permission string to revoke.}
}
\description{
Revoke permissions from an address
}
