% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mc_run_stream_filter}
\alias{mc_run_stream_filter}
\title{Run an existing stream filter against an item}
\usage{
mc_run_stream_filter(conn, filter, tx, vout = NULL)
}
\description{
Run an existing stream filter against an item
}
