% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mc_run_tx_filter}
\alias{mc_run_tx_filter}
\title{Run an existing transaction filter against a transaction}
\usage{
mc_run_tx_filter(conn, filter, tx)
}
\description{
Run an existing transaction filter against a transaction
}
