% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{mc_send_asset}
\alias{mc_send_asset}
\title{Send asset to another address}
\usage{
mc_send_asset(conn, address, asset, quantity)
}
\arguments{
\item{conn}{Connection object.}

\item{address}{Recipient address.}

\item{asset}{Asset name or assetref.}

\item{quantity}{Amount to send.}
}
\description{
Send asset to another address
}
