% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payments.R
\name{mc_send_asset_from}
\alias{mc_send_asset_from}
\title{Send a single asset from a specific address}
\usage{
mc_send_asset_from(
  conn,
  from_address,
  to_address,
  asset,
  quantity,
  native_amount = 0,
  comment = "",
  comment_to = ""
)
}
\arguments{
\item{conn}{Connection object.}

\item{from_address}{Sender address.}

\item{to_address}{Recipient address.}

\item{asset}{Asset name, ref or txid.}

\item{quantity}{Amount of asset.}

\item{native_amount}{Optional native currency.}

\item{comment}{Optional comment.}

\item{comment_to}{Optional comment-to.}
}
\description{
Send a single asset from a specific address
}
