% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payments.R
\name{mc_send_from}
\alias{mc_send_from}
\title{Send payment from a specific address}
\usage{
mc_send_from(
  conn,
  from_address,
  to_address,
  amounts,
  comment = "",
  comment_to = ""
)
}
\arguments{
\item{conn}{Connection object.}

\item{from_address}{Sender address.}

\item{to_address}{Recipient address.}

\item{amounts}{Numeric (native) or list of assets/metadata.}

\item{comment}{Optional comment.}

\item{comment_to}{Optional comment-to.}
}
\description{
Send payment from a specific address
}
