% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payments.R
\name{mc_send_with_data}
\alias{mc_send_with_data}
\title{Send payment with inline metadata}
\usage{
mc_send_with_data(conn, address, amounts, data)
}
\arguments{
\item{conn}{Connection object.}

\item{address}{Recipient address.}

\item{amounts}{Numeric (native) or list of assets.}

\item{data}{Data string or list (will be converted to hex).}
}
\description{
Send payment with inline metadata
}
