% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payments.R
\name{mc_send_with_data_from}
\alias{mc_send_with_data_from}
\title{Send payment from specific address with metadata}
\usage{
mc_send_with_data_from(conn, from_address, to_address, amounts, data)
}
\arguments{
\item{conn}{Connection object.}

\item{from_address}{Sender address.}

\item{to_address}{Recipient address.}

\item{amounts}{Numeric (native) or list of assets.}

\item{data}{Data string or list (will be converted to hex).}
}
\description{
Send payment from specific address with metadata
}
