% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_runtime.R
\name{mc_set_last_block}
\alias{mc_set_last_block}
\title{Rewind the node's active chain}
\usage{
mc_set_last_block(conn, hash_or_height)
}
\arguments{
\item{conn}{Connection object.}

\item{hash_or_height}{Block hash or height to rewind to.}
}
\value{
The hash of the last block after the change.
}
\description{
Note: Must be called after mc_pause(conn, "incoming,mining").
}
