% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mc_set_variable_value}
\alias{mc_set_variable_value}
\title{Set the value of a variable}
\usage{
mc_set_variable_value(conn, variable, value = NULL)
}
\arguments{
\item{value}{Any valid JSON structure (list, string, number).}
}
\description{
Set the value of a variable
}
