% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{mc_sign_message}
\alias{mc_sign_message}
\title{Sign a message with a private key}
\usage{
mc_sign_message(conn, address_or_key, message)
}
\arguments{
\item{conn}{Connection object.}

\item{address_or_key}{A wallet address (must be in this wallet) or a private key (WIF).}

\item{message}{The text message to sign.}
}
\value{
A string containing the base64-encoded signature.
}
\description{
Returns a base64-encoded digital signature which proves that the message
was approved by the owner of the address or private key.
}
