% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mc_test_library}
\alias{mc_test_library}
\title{Manage testing of libraries and updates locally}
\usage{
mc_test_library(conn, library = NULL, updatename = NULL, js_code = NULL)
}
\description{
Manage testing of libraries and updates locally
}
