% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mc_test_stream_filter}
\alias{mc_test_stream_filter}
\title{Test a stream filter before creation}
\usage{
mc_test_stream_filter(conn, options, js_code, tx = NULL, vout = NULL)
}
\arguments{
\item{vout}{Optional output index.}
}
\description{
Test a stream filter before creation
}
