% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{mc_update}
\alias{mc_update}
\title{Update asset status (open/closed)}
\usage{
mc_update(conn, asset, params)
}
\arguments{
\item{conn}{A connection object created by \code{\link{mc_connect}}.}

\item{asset}{Character string. Asset name, reference, or issuance ID.}

\item{params}{A list of parameters to update, typically
\code{list(open = FALSE)} to close further issuances.}
}
\value{
A character string containing the transaction ID.
}
\description{
Changes the status of an asset (e.g., to open or close further issuances).
The asset must have been created with the ability to be updated.
}
\examples{
\dontrun{
# Close the asset "mycoin" to prevent further issuances
txid <- mc_update(conn, "mycoin", list(open = FALSE))
}

}
\seealso{
\code{\link{mc_update_from}} to update from a specific address.

Other asset management: 
\code{\link{mc_update_from}()}
}
\concept{asset management}
