% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{mc_update_from}
\alias{mc_update_from}
\title{Update asset status from specific address}
\usage{
mc_update_from(conn, from_address, asset, params)
}
\description{
Update asset status from specific address
}
