% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multichainr-package.R
\docType{package}
\name{multichainr-package}
\alias{multichainr}
\alias{multichainr-package}
\title{multichainr: R Interface to the 'MultiChain' Blockchain RPC API}
\description{
Provides a comprehensive R interface to the 'MultiChain' blockchain JSON-RPC API \url{https://www.multichain.com/developers/json-rpc-api/}. Allows users to manage blockchain nodes, create and subscribe to data streams, issue assets, and manage network permissions directly from the R console. Supports both local node management and remote server interaction.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datascienceadvice/multichainr}
  \item Report bugs at \url{https://github.com/datascienceadvice/multichainr/issues}
}

}
\author{
\strong{Maintainer}: Aliaksandr Martsinkevich \email{alex@capsula.by} (\href{https://orcid.org/0000-0003-3655-4489}{ORCID})

}
\keyword{internal}
