% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{print.multichain_conn}
\alias{print.multichain_conn}
\title{Print MultiChain connection}
\usage{
\method{print}{multichain_conn}(x, ...)
}
\arguments{
\item{x}{A multichain_conn object.}

\item{...}{Additional arguments passed to print.}
}
\description{
Print MultiChain connection
}
