% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addresses.R
\name{mc_add_multisig_address}
\alias{mc_add_multisig_address}
\title{Add a multi-signature address}
\usage{
mc_add_multisig_address(conn, n_required, keys)
}
\arguments{
\item{conn}{Connection object.}

\item{n_required}{Number of required signatures (integer).}

\item{keys}{A character vector of public keys or addresses.}
}
\value{
A string containing the multisig address.
}
\description{
Creates a multi-signature address from a list of keys and required signatures.
}
