% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_transactions.R
\name{mc_append_raw_data}
\alias{mc_append_raw_data}
\title{Add metadata to a raw transaction}
\usage{
mc_append_raw_data(conn, tx_hex, data)
}
\arguments{
\item{conn}{Connection object.}

\item{tx_hex}{Raw transaction hex.}

\item{data}{Data string or list.}
}
\description{
Add metadata to a raw transaction
}
