% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomic_exchange.R
\name{mc_append_raw_exchange}
\alias{mc_append_raw_exchange}
\title{Add to a raw atomic exchange transaction}
\usage{
mc_append_raw_exchange(conn, tx_hex, txid, vout, amounts)
}
\arguments{
\item{conn}{Connection object.}

\item{tx_hex}{Hex string of the partial exchange transaction.}

\item{txid}{Transaction ID of the output being added to the offer.}

\item{vout}{Vout of the output being added.}

\item{amounts}{Assets/currency asked for in exchange for this addition.}
}
\value{
A list containing 'hex' (new partial transaction) and 'complete' (logical).
}
\description{
Add to a raw atomic exchange transaction
}
