% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_transactions.R
\name{mc_append_raw_transaction}
\alias{mc_append_raw_transaction}
\title{Add inputs and outputs to a raw transaction}
\usage{
mc_append_raw_transaction(conn, tx_hex, inputs = list(), outputs = list())
}
\arguments{
\item{conn}{A connection object created by \code{\link{mc_connect}}.}

\item{tx_hex}{Character string. The raw transaction hex to which inputs/outputs are added.}

\item{inputs}{A list of input objects (each with \code{txid} and \code{vout}).}

\item{outputs}{A named list of outputs (mapping addresses to amounts).}
}
\value{
A character string containing the updated raw transaction hex.
}
\description{
Appends additional inputs and outputs to an existing raw transaction.
This is useful for building multi‑party transactions or adding extra
components after creation.
}
\examples{
\dontrun{
# Add another input and output
new_input <- list(list(txid = "def...", vout = 1))
new_output <- list("1C..." = 0.2)
updated_hex <- mc_append_raw_transaction(conn, tx_hex,
                                         inputs = new_input,
                                         outputs = new_output)
}

}
\seealso{
\code{\link{mc_create_raw_transaction}}, \code{\link{mc_append_raw_change}}

Other raw transactions: 
\code{\link{mc_append_raw_change}()},
\code{\link{mc_append_raw_data}()},
\code{\link{mc_create_raw_send_from}()},
\code{\link{mc_create_raw_transaction}()},
\code{\link{mc_decode_raw_transaction}()},
\code{\link{mc_send_raw_transaction}()},
\code{\link{mc_sign_raw_transaction}()}
}
\concept{raw transactions}
