% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mc_approve_from}
\alias{mc_approve_from}
\title{Approve or disapprove an upgrade or filter}
\usage{
mc_approve_from(conn, from_address, entity, approve)
}
\arguments{
\item{conn}{Connection object.}

\item{from_address}{Admin address.}

\item{entity}{Name or txid of the upgrade/filter.}

\item{approve}{Boolean (for global) or list(for = "stream", approve = TRUE) for stream filters.}
}
\description{
Approve or disapprove an upgrade or filter
}
