% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet.R
\name{mc_backup_wallet}
\alias{mc_backup_wallet}
\title{Backup the wallet file}
\usage{
mc_backup_wallet(conn, filename)
}
\arguments{
\item{conn}{A connection object to the MultiChain node.}

\item{filename}{Character. The full path and filename for the backup.
\strong{Note:} This path is relative to the machine where the MultiChain node is running.}
}
\value{
Returns \code{NULL} on success.
}
\description{
Safely copies the \code{wallet.dat} file to a specified destination.
}
\seealso{
Other wallet: 
\code{\link{mc_change_wallet_passphrase}()},
\code{\link{mc_combine_unspent}()},
\code{\link{mc_dump_privkey}()},
\code{\link{mc_dump_wallet}()},
\code{\link{mc_encrypt_wallet}()},
\code{\link{mc_get_wallet_info}()},
\code{\link{mc_import_privkey}()},
\code{\link{mc_import_wallet}()},
\code{\link{mc_list_lock_unspent}()},
\code{\link{mc_list_unspent}()},
\code{\link{mc_lock_unspent}()},
\code{\link{mc_lock_wallet}()},
\code{\link{mc_unlock_wallet}()}
}
\concept{wallet}
