% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet.R
\name{mc_backup_wallet}
\alias{mc_backup_wallet}
\title{Backup the wallet file}
\usage{
mc_backup_wallet(conn, filename)
}
\arguments{
\item{conn}{Connection object.}

\item{filename}{Full path and name for the backup file on the node's machine.}
}
\description{
Backup the wallet file
}
