% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet.R
\name{mc_change_wallet_passphrase}
\alias{mc_change_wallet_passphrase}
\title{Change the wallet passphrase}
\usage{
mc_change_wallet_passphrase(conn, old_passphrase, new_passphrase)
}
\description{
Change the wallet passphrase
}
