% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_runtime.R
\name{mc_clear_mempool}
\alias{mc_clear_mempool}
\title{Clear the node's memory pool}
\usage{
mc_clear_mempool(conn)
}
\arguments{
\item{conn}{Connection object.}
}
\value{
NULL on success.
}
\description{
Removes all unconfirmed transactions from the mempool.
Note: Must be called after mc_pause(conn, "incoming,mining").
}
