% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet.R
\name{mc_combine_unspent}
\alias{mc_combine_unspent}
\title{Combine unspent outputs}
\usage{
mc_combine_unspent(
  conn,
  addresses = "*",
  minconf = 1,
  maxcombines = 100,
  mininputs = 2,
  maxinputs = 100,
  maxtime = 15
)
}
\arguments{
\item{conn}{Connection object.}

\item{addresses}{Vector of addresses or "\emph{" (default "}").}

\item{minconf}{Minimum confirmations (default 1).}

\item{maxcombines}{Maximum transactions to create (default 100).}

\item{mininputs}{Minimum inputs per transaction (default 2).}

\item{maxinputs}{Maximum inputs per transaction (default 100).}

\item{maxtime}{Maximum seconds to spend combining (default 15).}
}
\value{
A character vector of transaction IDs created.
}
\description{
Sends transactions to combine many small UTXOs into one to improve wallet performance.
}
