% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet.R
\name{mc_combine_unspent}
\alias{mc_combine_unspent}
\title{Combine unspent outputs (UTXOs)}
\usage{
mc_combine_unspent(
  conn,
  addresses = "*",
  minconf = 1,
  maxcombines = 100,
  mininputs = 2,
  maxinputs = 100,
  maxtime = 15
)
}
\arguments{
\item{conn}{A connection object to the MultiChain node.}

\item{addresses}{A vector of addresses, or \code{"*"} for all addresses (default \code{"*"}).}

\item{minconf}{Integer. Minimum confirmations (default \code{1}).}

\item{maxcombines}{Integer. Maximum number of transactions to create (default \code{100}).}

\item{mininputs}{Integer. Minimum number of inputs per transaction (default \code{2}).}

\item{maxinputs}{Integer. Maximum number of inputs per transaction (default \code{100}).}

\item{maxtime}{Integer. Maximum seconds to spend combining (default \code{15}).}
}
\value{
A character vector of the transaction IDs (txids) created.
}
\description{
Sends transactions to combine many small unspent transaction outputs (UTXOs)
into a single output. This is used to improve wallet performance and reduce
the size of the wallet's UTXO set.
}
\seealso{
Other wallet: 
\code{\link{mc_backup_wallet}()},
\code{\link{mc_change_wallet_passphrase}()},
\code{\link{mc_dump_privkey}()},
\code{\link{mc_dump_wallet}()},
\code{\link{mc_encrypt_wallet}()},
\code{\link{mc_get_wallet_info}()},
\code{\link{mc_import_privkey}()},
\code{\link{mc_import_wallet}()},
\code{\link{mc_list_lock_unspent}()},
\code{\link{mc_list_unspent}()},
\code{\link{mc_lock_unspent}()},
\code{\link{mc_lock_wallet}()},
\code{\link{mc_unlock_wallet}()}
}
\concept{wallet}
