% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomic_exchange.R
\name{mc_complete_raw_exchange}
\alias{mc_complete_raw_exchange}
\title{Finalize an atomic exchange transaction}
\usage{
mc_complete_raw_exchange(conn, tx_hex, txid, vout, amounts, data = NULL)
}
\arguments{
\item{conn}{Connection object.}

\item{tx_hex}{Hex string of the partial exchange transaction.}

\item{txid}{Transaction ID of the completing output.}

\item{vout}{Vout of the completing output.}

\item{amounts}{Assets/currency for the final part of the exchange.}

\item{data}{Optional metadata (string or list) to include.}
}
\value{
A raw transaction hex ready for sendrawtransaction.
}
\description{
Finalize an atomic exchange transaction
}
