% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_cache.R
\name{mc_create_binary_cache}
\alias{mc_create_binary_cache}
\title{Create a new binary cache item}
\usage{
mc_create_binary_cache(conn)
}
\arguments{
\item{conn}{Connection object.}
}
\value{
A string identifier (filename).
}
\description{
Creates an empty item (file) in the node's binary cache and returns its identifier.
}
