% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mc_create_library}
\alias{mc_create_library}
\title{Create a new library}
\usage{
mc_create_library(
  conn,
  name,
  updatemode = c("none", "instant", "approve"),
  js_code
)
}
\arguments{
\item{updatemode}{One of "none", "instant", or "approve".}
}
\description{
Create a new library
}
