% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomic_exchange.R
\name{mc_create_raw_exchange}
\alias{mc_create_raw_exchange}
\title{Create a new atomic exchange transaction}
\usage{
mc_create_raw_exchange(conn, txid, vout, amounts)
}
\arguments{
\item{conn}{Connection object.}

\item{txid}{Transaction ID from preparelockunspent.}

\item{vout}{Vout from preparelockunspent.}

\item{amounts}{Assets/currency asked for in exchange.}
}
\value{
A raw partial transaction in hexadecimal.
}
\description{
Create a new atomic exchange transaction
}
