% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_transactions.R
\name{mc_create_raw_send_from}
\alias{mc_create_raw_send_from}
\title{Create and fund a raw transaction from a specific address}
\usage{
mc_create_raw_send_from(
  conn,
  from_address,
  to_amounts,
  data = list(),
  action = ""
)
}
\arguments{
\item{conn}{Connection object.}

\item{from_address}{Sender address.}

\item{to_amounts}{List mapping recipient addresses to amounts.}

\item{data}{Optional metadata array.}

\item{action}{Optional action (e.g. "send", "sign").}
}
\description{
Create and fund a raw transaction from a specific address
}
