% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_transactions.R
\name{mc_create_raw_transaction}
\alias{mc_create_raw_transaction}
\title{Create a raw transaction}
\usage{
mc_create_raw_transaction(conn, inputs, outputs, data = list(), action = "")
}
\arguments{
\item{conn}{Connection object.}

\item{inputs}{List of lists, each containing "txid" and "vout".}

\item{outputs}{List/object mapping addresses to amounts.}

\item{data}{Optional array of metadata (strings or lists).}

\item{action}{Optional action: "lock", "sign", "lock,sign", or "send".}
}
\description{
Create a raw transaction
}
