% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{mc_create_stream}
\alias{mc_create_stream}
\title{Create a new stream}
\usage{
mc_create_stream(conn, name, open = TRUE, custom_fields = NULL)
}
\arguments{
\item{conn}{Connection object.}

\item{name}{Stream name.}

\item{open}{Boolean (TRUE for open stream) or a list of restrictions (e.g., list(restrict = "write")).}

\item{custom_fields}{Optional list of custom fields.}
}
\value{
The txid of the stream creation.
}
\description{
Create a new stream
}
