% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{mc_create_stream}
\alias{mc_create_stream}
\title{Create new stream}
\usage{
mc_create_stream(conn, name, open = TRUE)
}
\arguments{
\item{conn}{Connection object.}

\item{name}{Stream name.}

\item{open}{Boolean. If TRUE, anyone with "send" metadata permissions can publish.}
}
\description{
Create new stream
}
