% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{mc_create_stream_from}
\alias{mc_create_stream_from}
\title{Create a new stream from a specific address}
\usage{
mc_create_stream_from(
  conn,
  from_address,
  name,
  open = TRUE,
  custom_fields = NULL
)
}
\arguments{
\item{conn}{A connection object created by \code{\link{mc_connect}}.}

\item{from_address}{Character string. The address that will create the stream.}

\item{name}{Character string. Stream name.}

\item{open}{Either logical or a list of parameters (see \code{\link{mc_create_stream}}).}

\item{custom_fields}{Optional list of custom fields.}
}
\value{
A character string containing the transaction ID.
}
\description{
Creates a stream from a specified address (the address pays for the transaction).
This is useful when the node has multiple addresses and you want to control
which address appears as the creator.
}
\examples{
\dontrun{
txid <- mc_create_stream_from(conn, "1A...", "mystream", open = TRUE)
}

}
\seealso{
\code{\link{mc_create_stream}}

Other streams: 
\code{\link{mc_create_stream}()},
\code{\link{mc_get_stream_info}()},
\code{\link{mc_get_stream_item}()},
\code{\link{mc_get_stream_key_summary}()},
\code{\link{mc_get_stream_publisher_summary}()},
\code{\link{mc_list_stream_block_items}()},
\code{\link{mc_list_stream_items}()},
\code{\link{mc_list_stream_key_items}()},
\code{\link{mc_list_stream_keys}()},
\code{\link{mc_list_stream_publisher_items}()},
\code{\link{mc_list_stream_publishers}()},
\code{\link{mc_list_stream_query_items}()},
\code{\link{mc_list_stream_tx_items}()},
\code{\link{mc_list_streams}()},
\code{\link{mc_publish}()},
\code{\link{mc_publish_from}()},
\code{\link{mc_publish_multi}()},
\code{\link{mc_publish_multi_from}()}
}
\concept{streams}
