% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mc_create_tx_filter}
\alias{mc_create_tx_filter}
\title{Create a transaction filter}
\usage{
mc_create_tx_filter(conn, name, options, js_code)
}
\arguments{
\item{conn}{Connection object.}

\item{name}{Filter name.}

\item{options}{List of options, e.g., list(for = "asset1", libraries = list("lib1")).}

\item{js_code}{JavaScript code for the filter.}
}
\description{
Create a transaction filter
}
