% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mc_create_upgrade}
\alias{mc_create_upgrade}
\title{Create a blockchain upgrade}
\usage{
mc_create_upgrade(conn, name, params)
}
\arguments{
\item{conn}{Connection object.}

\item{name}{Upgrade name.}

\item{params}{List of parameters to upgrade (e.g., list("target-block-time" = 20)).}
}
\description{
Create a blockchain upgrade
}
