% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mc_create_variable}
\alias{mc_create_variable}
\title{Create a new variable}
\usage{
mc_create_variable(conn, name, open = TRUE, value = NULL)
}
\arguments{
\item{open}{If TRUE, anyone with create permissions can edit.}

\item{value}{Initial JSON value.}
}
\description{
Create a new variable
}
