% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_cache.R
\name{mc_delete_binary_cache}
\alias{mc_delete_binary_cache}
\title{Delete an item from the binary cache}
\usage{
mc_delete_binary_cache(conn, identifier)
}
\arguments{
\item{conn}{Connection object.}

\item{identifier}{The cache item identifier to remove.}
}
\value{
NULL on success.
}
\description{
Delete an item from the binary cache
}
