% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet.R
\name{mc_dump_privkey}
\alias{mc_dump_privkey}
\title{Dump a private key for an address}
\usage{
mc_dump_privkey(conn, address)
}
\arguments{
\item{conn}{Connection object.}

\item{address}{The wallet address.}
}
\value{
The private key in WIF format.
}
\description{
Dump a private key for an address
}
