% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet.R
\name{mc_dump_wallet}
\alias{mc_dump_wallet}
\title{Dump all private keys to a file}
\usage{
mc_dump_wallet(conn, filename)
}
\arguments{
\item{conn}{Connection object.}

\item{filename}{Full path for the text file on the node's machine.}
}
\description{
Dump all private keys to a file
}
