% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet.R
\name{mc_encrypt_wallet}
\alias{mc_encrypt_wallet}
\title{Encrypt the wallet for the first time}
\usage{
mc_encrypt_wallet(conn, passphrase)
}
\arguments{
\item{conn}{Connection object.}

\item{passphrase}{The new wallet password.}
}
\description{
WARNING: MultiChain will stop after this and must be restarted.
}
