% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{mc_get_address_balances}
\alias{mc_get_address_balances}
\title{Get asset balances for an address}
\usage{
mc_get_address_balances(conn, address)
}
\arguments{
\item{conn}{Connection object.}

\item{address}{Wallet address.}
}
\description{
Get asset balances for an address
}
