% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{mc_get_address_balances}
\alias{mc_get_address_balances}
\title{Get asset balances for a specific address}
\usage{
mc_get_address_balances(conn, address, minconf = 1, include_locked = FALSE)
}
\arguments{
\item{conn}{A connection object to the MultiChain node.}

\item{address}{Character. The MultiChain address to query.}

\item{minconf}{Integer. The minimum number of confirmations (default \code{1}).}

\item{include_locked}{Logical. If \code{TRUE}, includes unspent outputs that have been locked (default \code{FALSE}).}
}
\value{
A data frame containing asset names, amounts, and other balance details.
}
\description{
Returns a list of all asset balances for a given address in the node's wallet.
}
\seealso{
Other transactions: 
\code{\link{mc_get_address_transaction}()},
\code{\link{mc_get_multi_balances}()},
\code{\link{mc_get_token_balances}()},
\code{\link{mc_get_total_balances}()},
\code{\link{mc_get_tx_out_data}()},
\code{\link{mc_get_wallet_transaction}()},
\code{\link{mc_list_address_transactions}()},
\code{\link{mc_list_wallet_transactions}()},
\code{\link{mc_send}()},
\code{\link{mc_send_asset}()},
\code{\link{mc_send_asset_from}()},
\code{\link{mc_send_from}()},
\code{\link{mc_send_with_data}()},
\code{\link{mc_send_with_data_from}()}
}
\concept{transactions}
