% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{mc_get_address_transaction}
\alias{mc_get_address_transaction}
\title{Get details of a transaction for a specific address}
\usage{
mc_get_address_transaction(conn, address, txid, verbose = FALSE)
}
\arguments{
\item{conn}{A connection object to the MultiChain node.}

\item{address}{Character. The MultiChain address to query.}

\item{txid}{Character. The transaction ID.}

\item{verbose}{Logical. If \code{TRUE}, provides a detailed breakdown of inputs and outputs (default \code{FALSE}).}
}
\value{
A list containing transaction details.
}
\description{
Provides information about a specific transaction, but only if it involves the specified address.
}
\seealso{
Other MultiChain Transactions: 
\code{\link{mc_get_tx_out_data}()},
\code{\link{mc_get_wallet_transaction}()},
\code{\link{mc_list_address_transactions}()},
\code{\link{mc_list_wallet_transactions}()}
}
\concept{MultiChain Transactions}
