% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{mc_get_asset_transaction}
\alias{mc_get_asset_transaction}
\title{Get a specific transaction involving a subscribed asset}
\usage{
mc_get_asset_transaction(conn, asset, txid, verbose = FALSE)
}
\description{
Get a specific transaction involving a subscribed asset
}
