% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_local.R
\name{mc_get_bin_path}
\alias{mc_get_bin_path}
\title{Internal function to locate executable files}
\usage{
mc_get_bin_path(bin_name)
}
\arguments{
\item{bin_name}{Name of the binary (multichaind or multichain-util).}
}
\description{
Internal function to locate executable files
}
\keyword{internal}
