% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockchain.R
\name{mc_get_block}
\alias{mc_get_block}
\title{Get block information}
\usage{
mc_get_block(conn, hash_or_height, verbose = 1)
}
\arguments{
\item{conn}{Connection object.}

\item{hash_or_height}{Block hash or height.}

\item{verbose}{Verbosity level (0 to 4). Default 1.}
}
\description{
Get block information
}
