% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{mc_get_blockchain_params}
\alias{mc_get_blockchain_params}
\title{Get blockchain parameters}
\usage{
mc_get_blockchain_params(conn)
}
\arguments{
\item{conn}{A multichain_conn object.}
}
\value{
A list containing blockchain configuration parameters (e.g., protocol version,
target block time, maximum block size, etc.).
}
\description{
Returns an object containing the parameters used to initialize this blockchain.
}
