% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_runtime.R
\name{mc_get_chunk_queue_info}
\alias{mc_get_chunk_queue_info}
\title{Get information about off-chain chunk queue}
\usage{
mc_get_chunk_queue_info(conn)
}
\arguments{
\item{conn}{A connection object created by \code{\link{mc_connect}}.}
}
\value{
A list containing:
\item{chunk_count}{Number of chunks currently queued.}
\item{byte_count}{Total size in bytes of queued chunks.}
\item{...}{Other queue statistics.}
}
\description{
Returns details about the node's off‑chain chunk queue, which handles
the transmission of large data items that are split into chunks.
}
\examples{
\dontrun{
queue_info <- mc_get_chunk_queue_info(conn)
cat("Chunks pending:", queue_info$chunk_count)
}

}
\seealso{
\code{\link{mc_get_chunk_queue_totals}} for cumulative statistics.

Other off-chain data: 
\code{\link{mc_get_chunk_queue_totals}()}
}
\concept{off-chain data}
