% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_management.R
\name{mc_get_config}
\alias{mc_get_config}
\title{Get MultiChain config}
\usage{
mc_get_config(chain_name, base_dir = NULL)
}
\arguments{
\item{chain_name}{Name of the chain.}

\item{base_dir}{Optional base directory path (for testing).}
}
\description{
Get MultiChain config
}
