% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{mc_get_info}
\alias{mc_get_info}
\title{Get general node information}
\usage{
mc_get_info(conn)
}
\arguments{
\item{conn}{A multichain_conn object.}
}
\value{
A list containing node parameters (version, protocol, balance, etc.).
}
\description{
Returns an object containing information about the status of the node,
network, and blockchain.
}
