% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{mc_get_init_status}
\alias{mc_get_init_status}
\title{Get node initialization status}
\usage{
mc_get_init_status(conn)
}
\arguments{
\item{conn}{A multichain_conn object.}
}
\value{
A list containing status (string) and progress (numeric, 0 to 1).
}
\description{
Returns information about the node's initialization status,
which is especially useful during startup.
}
