% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockchain.R
\name{mc_get_last_block_info}
\alias{mc_get_last_block_info}
\title{Get information about the last block}
\usage{
mc_get_last_block_info(conn, skip = 0)
}
\arguments{
\item{skip}{Number of blocks to skip back (default 0).}
}
\description{
Get information about the last block
}
