% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{mc_get_multi_balances}
\alias{mc_get_multi_balances}
\title{Get balances for multiple addresses and assets}
\usage{
mc_get_multi_balances(
  conn,
  addresses = "*",
  assets = "*",
  minconf = 1,
  include_watch_only = FALSE,
  include_locked = FALSE
)
}
\arguments{
\item{conn}{Connection object.}

\item{addresses}{Vector of addresses or "\emph{" for all (default "}").}

\item{assets}{Vector of asset names/refs or "\emph{" for all (default "}").}

\item{minconf}{Minimum confirmations (default 1).}

\item{include_watch_only}{Include watch-only addresses (default FALSE).}

\item{include_locked}{Include locked unspent outputs (default FALSE).}
}
\description{
Get balances for multiple addresses and assets
}
