% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{mc_get_network_info}
\alias{mc_get_network_info}
\title{Get information about node's network status}
\usage{
mc_get_network_info(conn)
}
\arguments{
\item{conn}{Connection object.}
}
\value{
A list containing port and local address information.
}
\description{
Get information about node's network status
}
