% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockchain.R
\name{mc_get_raw_transaction}
\alias{mc_get_raw_transaction}
\title{Get a raw transaction from the blockchain}
\usage{
mc_get_raw_transaction(conn, txid, verbose = FALSE)
}
\arguments{
\item{txid}{Transaction ID.}

\item{verbose}{If TRUE, returns a detailed list.}
}
\description{
Get a raw transaction from the blockchain
}
