% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{mc_get_runtime_params}
\alias{mc_get_runtime_params}
\title{Get node runtime parameters}
\usage{
mc_get_runtime_params(conn)
}
\arguments{
\item{conn}{A multichain_conn object.}
}
\value{
A list of runtime parameters (e.g., mining status, connection limits, etc.).
}
\description{
Returns an object containing the current runtime parameters of the node.
}
