% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{mc_get_stream_info}
\alias{mc_get_stream_info}
\title{Get information about a specific stream}
\usage{
mc_get_stream_info(conn, stream, verbose = FALSE)
}
\arguments{
\item{conn}{Connection object.}

\item{stream}{Stream name, ref or creation txid.}

\item{verbose}{If TRUE, includes creator information.}
}
\description{
Get information about a specific stream
}
