% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{mc_get_stream_key_summary}
\alias{mc_get_stream_key_summary}
\title{Get summary of JSON objects in a stream for a specific key}
\usage{
mc_get_stream_key_summary(conn, stream, key, mode = "jsonobjectmerge")
}
\arguments{
\item{mode}{Default "jsonobjectmerge". Can include "recursive", "noupdate", etc.}
}
\description{
Get summary of JSON objects in a stream for a specific key
}
