% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{mc_get_total_balances}
\alias{mc_get_total_balances}
\title{Get total wallet balances}
\usage{
mc_get_total_balances(
  conn,
  minconf = 1,
  include_watch_only = FALSE,
  include_locked = FALSE
)
}
\arguments{
\item{conn}{Connection object.}

\item{minconf}{Minimum confirmations (default 1).}

\item{include_watch_only}{Include watch-only addresses (default FALSE).}

\item{include_locked}{Include locked outputs (default FALSE).}
}
\description{
Get total wallet balances
}
