% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockchain.R
\name{mc_get_tx_out}
\alias{mc_get_tx_out}
\title{Get details about an unspent transaction output}
\usage{
mc_get_tx_out(conn, txid, vout, unconfirmed = FALSE)
}
\arguments{
\item{txid}{Transaction ID.}

\item{vout}{Output index.}

\item{unconfirmed}{Include unconfirmed transactions.}
}
\description{
Get details about an unspent transaction output
}
