% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{mc_get_tx_out_data}
\alias{mc_get_tx_out_data}
\title{Retrieve full hex data from a transaction output}
\usage{
mc_get_tx_out_data(conn, txid, vout, count_bytes = NULL, start_byte = 0)
}
\arguments{
\item{conn}{A connection object to the MultiChain node.}

\item{txid}{Character. The transaction ID containing the output.}

\item{vout}{Integer. The index of the output (starting from 0).}

\item{count_bytes}{Integer. The number of bytes to retrieve. If \code{NULL}, returns all data.}

\item{start_byte}{Integer. The byte offset to start reading from (default \code{0}).}
}
\value{
A list or string containing the hex data from the transaction output.
}
\description{
This function is used to retrieve large data payloads (like stream data) that
may have been truncated in standard transaction calls.
}
\seealso{
Other transactions: 
\code{\link{mc_get_address_balances}()},
\code{\link{mc_get_address_transaction}()},
\code{\link{mc_get_multi_balances}()},
\code{\link{mc_get_token_balances}()},
\code{\link{mc_get_total_balances}()},
\code{\link{mc_get_wallet_transaction}()},
\code{\link{mc_list_address_transactions}()},
\code{\link{mc_list_wallet_transactions}()},
\code{\link{mc_send}()},
\code{\link{mc_send_asset}()},
\code{\link{mc_send_asset_from}()},
\code{\link{mc_send_from}()},
\code{\link{mc_send_with_data}()},
\code{\link{mc_send_with_data_from}()}
}
\concept{transactions}
