% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mc_get_variable_history}
\alias{mc_get_variable_history}
\title{List historical values of a variable}
\usage{
mc_get_variable_history(
  conn,
  variable,
  verbose = FALSE,
  count = 10,
  start = NULL
)
}
\description{
List historical values of a variable
}
