% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{mc_get_wallet_transaction}
\alias{mc_get_wallet_transaction}
\title{Get details of a wallet transaction}
\usage{
mc_get_wallet_transaction(
  conn,
  txid,
  include_watch_only = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{conn}{A connection object to the MultiChain node.}

\item{txid}{Character. The transaction ID.}

\item{include_watch_only}{Logical. Include watch-only addresses (default \code{FALSE}).}

\item{verbose}{Logical. If \code{TRUE}, provides details of inputs and outputs (default \code{FALSE}).}
}
\value{
A list containing detailed transaction data.
}
\description{
Retrieves detailed information about a transaction in the node's wallet.
}
\seealso{
Other MultiChain Transactions: 
\code{\link{mc_get_address_transaction}()},
\code{\link{mc_get_tx_out_data}()},
\code{\link{mc_list_address_transactions}()},
\code{\link{mc_list_wallet_transactions}()}
}
\concept{MultiChain Transactions}
