% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{mc_grant_with_data_from}
\alias{mc_grant_with_data_from}
\title{Grant permissions from a specific address with metadata}
\usage{
mc_grant_with_data_from(
  conn,
  from_address,
  to_address,
  permissions,
  data,
  native_amount = 0
)
}
\arguments{
\item{conn}{A multichain_conn object.}

\item{from_address}{Address from which the permissions are granted.}

\item{to_address}{Target wallet address.}

\item{permissions}{Permission string.}

\item{data}{Data string or list (will be converted to hex).}

\item{native_amount}{Optional native currency amount.}
}
\description{
Grant permissions from a specific address with metadata
}
