% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{mc_help}
\alias{mc_help}
\title{Get help for MultiChain commands}
\usage{
mc_help(conn, command = NULL)
}
\arguments{
\item{conn}{A multichain_conn object.}

\item{command}{Optional string. The name of the command to get help for.}
}
\value{
An object of class \code{mc_help} containing the help text.
}
\description{
Returns a list of all available commands, or detailed help for a specific command.
}
