% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet.R
\name{mc_import_wallet}
\alias{mc_import_wallet}
\title{Import keys from a wallet dump file}
\usage{
mc_import_wallet(conn, filename, rescan = 0)
}
\arguments{
\item{conn}{A connection object to the MultiChain node.}

\item{filename}{Character. Path to the dump file on the node's machine.}

\item{rescan}{Integer. The block number to start rescanning from (default \code{0}).}
}
\value{
Returns \code{NULL} on success.
}
\description{
Import keys from a wallet dump file
}
\seealso{
Other MultiChain Wallet Management: 
\code{\link{mc_backup_wallet}()},
\code{\link{mc_change_wallet_passphrase}()},
\code{\link{mc_dump_privkey}()},
\code{\link{mc_dump_wallet}()},
\code{\link{mc_encrypt_wallet}()},
\code{\link{mc_get_wallet_info}()},
\code{\link{mc_import_privkey}()},
\code{\link{mc_lock_wallet}()},
\code{\link{mc_unlock_wallet}()}
}
\concept{MultiChain Wallet Management}
