% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{mc_issue}
\alias{mc_issue}
\title{Issue new asset}
\usage{
mc_issue(conn, address, name, quantity, units = 1)
}
\arguments{
\item{conn}{Connection object.}

\item{address}{Wallet address to receive the issued assets.}

\item{name}{Asset name string or list of parameters.}

\item{quantity}{Total amount to issue.}

\item{units}{Smallest divisible unit (e.g., 0.01).}
}
\description{
Issue new asset
}
