% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{mc_issue_more}
\alias{mc_issue_more}
\title{Issue more of an existing fungible asset}
\usage{
mc_issue_more(
  conn,
  address,
  asset,
  quantity,
  native_amount = NULL,
  custom_fields = NULL
)
}
\arguments{
\item{conn}{Connection object.}

\item{address}{Recipient address.}

\item{asset}{Asset name, ref or txid.}

\item{quantity}{Amount to issue.}

\item{native_amount}{Optional native currency amount.}

\item{custom_fields}{Optional list of custom fields.}
}
\description{
Issue more of an existing fungible asset
}
