% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{mc_issue_token}
\alias{mc_issue_token}
\title{Issue tokens for a non-fungible asset (NFT)}
\usage{
mc_issue_token(
  conn,
  address,
  asset,
  token,
  quantity,
  native_amount = NULL,
  token_details = NULL
)
}
\arguments{
\item{conn}{Connection object.}

\item{address}{Recipient address.}

\item{asset}{Parent asset name, ref or txid.}

\item{token}{Token name.}

\item{quantity}{Amount of tokens.}

\item{native_amount}{Optional native currency amount.}

\item{token_details}{Optional list of token details.}
}
\description{
Issue tokens for a non-fungible asset (NFT)
}
