% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{mc_list_address_transactions}
\alias{mc_list_address_transactions}
\title{List transactions for a specific address}
\usage{
mc_list_address_transactions(
  conn,
  address,
  count = 10,
  skip = 0,
  verbose = FALSE
)
}
\arguments{
\item{conn}{A connection object to the MultiChain node.}

\item{address}{Character. The MultiChain address to query.}

\item{count}{Integer. The number of transactions to return (default \code{10}).}

\item{skip}{Integer. The number of transactions to skip (default \code{0}).}

\item{verbose}{Logical. If \code{TRUE}, provides details of inputs and outputs (default \code{FALSE}).}
}
\value{
A data frame of transaction history for the address.
}
\description{
Returns a list of the most recent transactions involving the specified address.
}
\seealso{
Other MultiChain Transactions: 
\code{\link{mc_get_address_transaction}()},
\code{\link{mc_get_tx_out_data}()},
\code{\link{mc_get_wallet_transaction}()},
\code{\link{mc_list_wallet_transactions}()}
}
\concept{MultiChain Transactions}
