% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addresses.R
\name{mc_list_addresses}
\alias{mc_list_addresses}
\title{List addresses in the wallet}
\usage{
mc_list_addresses(
  conn,
  addresses = "*",
  verbose = FALSE,
  count = NULL,
  start = NULL
)
}
\arguments{
\item{conn}{Connection object.}

\item{addresses}{A specific address or vector of addresses to filter. Default "*" (all).}

\item{verbose}{If TRUE, provides detailed information.}

\item{count}{Optional. Number of addresses to list.}

\item{start}{Optional. Offset for the list.}
}
\value{
A data frame containing address information.
}
\description{
Returns information about the addresses in the current node's wallet.
}
