% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet.R
\name{mc_list_lock_unspent}
\alias{mc_list_lock_unspent}
\title{List locked unspent outputs}
\usage{
mc_list_lock_unspent(conn)
}
\arguments{
\item{conn}{A connection object to the MultiChain node.}
}
\value{
A data frame containing columns for \code{txid} and \code{vout}.
}
\description{
Returns a list of unspent transaction outputs that have been temporarily
locked by \code{\link{mc_lock_unspent}}.
}
\seealso{
Other wallet: 
\code{\link{mc_backup_wallet}()},
\code{\link{mc_change_wallet_passphrase}()},
\code{\link{mc_combine_unspent}()},
\code{\link{mc_dump_privkey}()},
\code{\link{mc_dump_wallet}()},
\code{\link{mc_encrypt_wallet}()},
\code{\link{mc_get_wallet_info}()},
\code{\link{mc_import_privkey}()},
\code{\link{mc_import_wallet}()},
\code{\link{mc_list_unspent}()},
\code{\link{mc_lock_unspent}()},
\code{\link{mc_lock_wallet}()},
\code{\link{mc_unlock_wallet}()}
}
\concept{wallet}
