% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet.R
\name{mc_list_lock_unspent}
\alias{mc_list_lock_unspent}
\title{List locked unspent outputs}
\usage{
mc_list_lock_unspent(conn)
}
\arguments{
\item{conn}{A connection object to the MultiChain node.}
}
\value{
A data frame containing columns for \code{txid} and \code{vout}.
}
\description{
Returns a list of unspent transaction outputs that have been temporarily
locked by \code{\link{mc_lock_unspent}}.
}
\seealso{
Other MultiChain UTXO Management: 
\code{\link{mc_combine_unspent}()},
\code{\link{mc_list_unspent}()},
\code{\link{mc_lock_unspent}()}
}
\concept{MultiChain UTXO Management}
