% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{mc_list_permissions}
\alias{mc_list_permissions}
\title{List network permissions}
\usage{
mc_list_permissions(conn, permissions = "*")
}
\arguments{
\item{conn}{A multichain_conn object.}

\item{permissions}{Permission type to filter (default "*").}
}
\description{
List network permissions
}
