% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{mc_list_stream_block_items}
\alias{mc_list_stream_block_items}
\title{List items in a specific block or blocks}
\usage{
mc_list_stream_block_items(
  conn,
  stream,
  blocks,
  verbose = FALSE,
  count = NULL,
  start = NULL
)
}
\arguments{
\item{blocks}{Vector of heights, hashes or timestamps.}
}
\description{
List items in a specific block or blocks
}
