% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{mc_list_stream_items}
\alias{mc_list_stream_items}
\title{List items in a stream}
\usage{
mc_list_stream_items(conn, stream, count = 10)
}
\arguments{
\item{conn}{Connection object.}

\item{stream}{Stream name.}

\item{count}{Number of items to retrieve.}
}
\description{
List items in a stream
}
