% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{mc_list_stream_keys}
\alias{mc_list_stream_keys}
\title{List unique keys in a stream}
\usage{
mc_list_stream_keys(
  conn,
  stream,
  keys = "*",
  verbose = FALSE,
  count = NULL,
  start = NULL,
  local_ordering = FALSE
)
}
\description{
List unique keys in a stream
}
