% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{mc_list_stream_query_items}
\alias{mc_list_stream_query_items}
\title{Query items by matching keys and publishers}
\usage{
mc_list_stream_query_items(conn, stream, query, verbose = FALSE)
}
\arguments{
\item{query}{List like list(keys = c("k1"), publishers = c("addr1"))}
}
\description{
Query items by matching keys and publishers
}
