% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet.R
\name{mc_list_unspent}
\alias{mc_list_unspent}
\title{List unspent transaction outputs}
\usage{
mc_list_unspent(conn, minconf = 1, maxconf = 999999, addresses = NULL)
}
\arguments{
\item{conn}{Connection object.}

\item{minconf}{Minimum confirmations (default 1).}

\item{maxconf}{Maximum confirmations (default 999999).}

\item{addresses}{Optional vector of addresses to filter.}
}
\value{
A data frame containing UTXO details, including assets and permissions.
}
\description{
List unspent transaction outputs
}
