% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet.R
\name{mc_lock_unspent}
\alias{mc_lock_unspent}
\title{Lock or unlock unspent outputs}
\usage{
mc_lock_unspent(conn, unlock, outputs = NULL)
}
\arguments{
\item{conn}{Connection object.}

\item{unlock}{Logical. If TRUE, unlocks; if FALSE, locks.}

\item{outputs}{Optional list of outputs to (un)lock: list(list(txid="id", vout=0), ...).}
}
\value{
Logical TRUE on success.
}
\description{
Lock or unlock unspent outputs
}
