% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_management.R
\name{mc_node_stop}
\alias{mc_node_stop}
\title{Stop a MultiChain node}
\usage{
mc_node_stop(x)
}
\arguments{
\item{x}{Either a connection object or a chain name.}
}
\description{
Stop a MultiChain node
}
