% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_runtime.R
\name{mc_pause}
\alias{mc_pause}
\title{Pause specified node tasks}
\usage{
mc_pause(conn, tasks)
}
\arguments{
\item{conn}{Connection object.}

\item{tasks}{A character vector or comma-separated string of tasks
(mining, incoming, offchain).}
}
\value{
NULL on success.
}
\description{
Pause specified node tasks
}
