% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomic_exchange.R
\name{mc_prepare_lock_unspent}
\alias{mc_prepare_lock_unspent}
\title{Prepare an unspent transaction output for exchange}
\usage{
mc_prepare_lock_unspent(conn, amounts, lock = TRUE)
}
\arguments{
\item{conn}{Connection object.}

\item{amounts}{List of assets or numeric amount of native currency.}

\item{lock}{If TRUE (default), the output is locked for automatic spending.}
}
\value{
A list containing 'txid' and 'vout'.
}
\description{
Prepare an unspent transaction output for exchange
}
