% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{mc_publish}
\alias{mc_publish}
\title{Publish data to a stream}
\usage{
mc_publish(conn, stream, key, data)
}
\arguments{
\item{conn}{Connection object.}

\item{stream}{Stream name.}

\item{key}{Item key.}

\item{data}{Data string or list (will be converted to JSON).}
}
\description{
Publish data to a stream
}
