% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{mc_publish_from}
\alias{mc_publish_from}
\title{Publish an item to a stream from a specific address}
\usage{
mc_publish_from(conn, from_address, stream, keys, data, options = NULL)
}
\arguments{
\item{conn}{Connection object.}

\item{from_address}{Address used to publish the item.}

\item{stream}{Stream name, ref or txid.}

\item{keys}{A single key or vector of keys.}

\item{data}{Data string or list.}

\item{options}{Optional "offchain" string.}
}
\description{
Publish an item to a stream from a specific address
}
