% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{mc_publish_multi}
\alias{mc_publish_multi}
\title{Publish multiple items to a stream in one transaction}
\usage{
mc_publish_multi(conn, stream, items, options = NULL)
}
\arguments{
\item{conn}{Connection object.}

\item{stream}{Default stream name, ref or txid.}

\item{items}{A list of lists, each containing "key" (or "keys") and "data".}

\item{options}{Optional default "offchain" string.}
}
\description{
Publish multiple items to a stream in one transaction
}
