% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{mc_publish_multi_from}
\alias{mc_publish_multi_from}
\title{Publish multiple items from a specific address}
\usage{
mc_publish_multi_from(conn, from_address, stream, items, options = NULL)
}
\arguments{
\item{conn}{Connection object.}

\item{from_address}{Address used to publish the items.}

\item{stream}{Default stream.}

\item{items}{List of lists.}

\item{options}{Optional default "offchain" string.}
}
\description{
Publish multiple items from a specific address
}
