% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_runtime.R
\name{mc_resume}
\alias{mc_resume}
\title{Resume specified node tasks}
\usage{
mc_resume(conn, tasks)
}
\arguments{
\item{conn}{Connection object.}

\item{tasks}{A character vector or comma-separated string of tasks.}
}
\value{
NULL on success.
}
\description{
Resume specified node tasks
}
