% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mc_rpc}
\alias{mc_rpc}
\title{Internal function for JSON-RPC requests}
\usage{
mc_rpc(conn, method, params = list())
}
\arguments{
\item{conn}{A multichain_conn object.}

\item{method}{Character string of the RPC method.}

\item{params}{List of parameters.}
}
\value{
The result field from the JSON-RPC response.
}
\description{
Internal function for JSON-RPC requests
}
\keyword{internal}
