% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payments.R
\name{mc_send}
\alias{mc_send}
\title{Send payment or assets to an address}
\usage{
mc_send(conn, address, amounts, comment = "", comment_to = "")
}
\arguments{
\item{conn}{Connection object.}

\item{address}{Recipient address.}

\item{amounts}{Numeric (native currency) or a list of assets/metadata.}

\item{comment}{Optional transaction comment.}

\item{comment_to}{Optional comment-to.}
}
\description{
Send payment or assets to an address
}
