% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payments.R
\name{mc_send_asset}
\alias{mc_send_asset}
\title{Send a single asset to an address}
\usage{
mc_send_asset(
  conn,
  address,
  asset,
  quantity,
  native_amount = 0,
  comment = "",
  comment_to = ""
)
}
\arguments{
\item{conn}{Connection object.}

\item{address}{Recipient address.}

\item{asset}{Asset name, ref or txid.}

\item{quantity}{Amount of asset to send.}

\item{native_amount}{Optional native currency to send (default 0).}

\item{comment}{Optional comment.}

\item{comment_to}{Optional comment-to.}
}
\description{
Send a single asset to an address
}
