% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_transactions.R
\name{mc_send_raw_transaction}
\alias{mc_send_raw_transaction}
\title{Send a signed raw transaction to the network}
\usage{
mc_send_raw_transaction(conn, tx_hex)
}
\value{
Transaction ID.
}
\description{
Send a signed raw transaction to the network
}
