% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_management.R
\name{mc_set_path}
\alias{mc_set_path}
\title{Set path to MultiChain binaries}
\usage{
mc_set_path(path)
}
\arguments{
\item{path}{Path to the folder containing multichaind and multichain-util.}
}
\description{
Set path to MultiChain binaries
}
