% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{mc_set_runtime_param}
\alias{mc_set_runtime_param}
\title{Set node runtime parameter}
\usage{
mc_set_runtime_param(conn, name, value)
}
\arguments{
\item{conn}{A multichain_conn object.}

\item{name}{The name of the parameter to change. Supported: acceptfiltertimeout,
autosubscribe, bantx, handshakelocal, hideknownopdrops, lockadminminerounds,
lockblock, lockinlinemetadata, maxshowndata, maxqueryscanitems,
mineemptyrounds, miningrequirespeers, miningturnover, sendfiltertimeout.}

\item{value}{The new value for the parameter. Can be boolean, numeric, or string.}
}
\value{
NULL on success.
}
\description{
Changes a runtime parameter of the node without requiring a restart.
}
