% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mc_set_variable_value}
\alias{mc_set_variable_value}
\title{Set the value of a variable}
\usage{
mc_set_variable_value(conn, variable, value = NULL)
}
\arguments{
\item{conn}{A connection object created by \code{\link{mc_connect}}.}

\item{variable}{Character string. Variable name or transaction ID.}

\item{value}{Optional. New value (any JSON structure). If \code{NULL}, the
variable is deleted.}
}
\value{
A list containing the result of the RPC call (usually transaction ID).
}
\description{
Updates the value of an existing variable. The new value can be any JSON‑compatible
object (list, number, string, etc.).
}
\examples{
\dontrun{
mc_set_variable_value(conn, "myvar", value = list(updated = TRUE))
}

}
\seealso{
\code{\link{mc_create_variable}}, \code{\link{mc_get_variable_value}}

Other variables: 
\code{\link{mc_create_variable}()},
\code{\link{mc_create_variable_from}()},
\code{\link{mc_get_variable_history}()},
\code{\link{mc_get_variable_info}()},
\code{\link{mc_get_variable_value}()},
\code{\link{mc_list_variables}()},
\code{\link{mc_set_variable_value_from}()}
}
\concept{variables}
