% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{mc_store_node}
\alias{mc_store_node}
\title{Add an IP address to known peer nodes (MultiChain 2.3+)}
\usage{
mc_store_node(conn, node, command = c("tryconnect", "ignore"))
}
\arguments{
\item{conn}{Connection object.}

\item{node}{IP address and port.}

\item{command}{One of "tryconnect" (default) or "ignore".}
}
\description{
Add an IP address to known peer nodes (MultiChain 2.3+)
}
