% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\name{mc_subscribe}
\alias{mc_subscribe}
\title{Subscribe to assets or streams}
\usage{
mc_subscribe(conn, entities, rescan = TRUE)
}
\arguments{
\item{conn}{Connection object.}

\item{entities}{A single name/ref/txid or a vector of them.}

\item{rescan}{If TRUE (default), the node reindexes all items from the creation of the entities.}
}
\value{
NULL on success.
}
\description{
Instructs the node to start tracking one or more asset(s) or stream(s).
}
