% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{mc_subscribe}
\alias{mc_subscribe}
\title{Subscribe to a stream}
\usage{
mc_subscribe(conn, stream)
}
\arguments{
\item{conn}{Connection object.}

\item{stream}{Stream name.}
}
\description{
Node must be subscribed to a stream to list its items.
}
