% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mc_test_tx_filter}
\alias{mc_test_tx_filter}
\title{Test a transaction filter before creation}
\usage{
mc_test_tx_filter(conn, options, js_code, tx = NULL)
}
\arguments{
\item{options}{Filter options.}

\item{js_code}{JavaScript code.}

\item{tx}{Hex or txid of a transaction to test against.}
}
\description{
Test a transaction filter before creation
}
