% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet.R
\name{mc_unlock_wallet}
\alias{mc_unlock_wallet}
\title{Unlock the wallet with a passphrase}
\usage{
mc_unlock_wallet(conn, passphrase, timeout)
}
\arguments{
\item{conn}{A connection object to the MultiChain node.}

\item{passphrase}{Character. The wallet password.}

\item{timeout}{Integer. Time in seconds to keep the wallet unlocked.}
}
\value{
Returns \code{NULL} on success.
}
\description{
Stores the wallet decryption key in memory for a specified duration.
}
\seealso{
Other MultiChain Wallet Management: 
\code{\link{mc_backup_wallet}()},
\code{\link{mc_change_wallet_passphrase}()},
\code{\link{mc_dump_privkey}()},
\code{\link{mc_dump_wallet}()},
\code{\link{mc_encrypt_wallet}()},
\code{\link{mc_get_wallet_info}()},
\code{\link{mc_import_privkey}()},
\code{\link{mc_import_wallet}()},
\code{\link{mc_lock_wallet}()}
}
\concept{MultiChain Wallet Management}
