% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet.R
\name{mc_unlock_wallet}
\alias{mc_unlock_wallet}
\title{Unlock the wallet with a passphrase}
\usage{
mc_unlock_wallet(conn, passphrase, timeout)
}
\arguments{
\item{conn}{Connection object.}

\item{passphrase}{The wallet password.}

\item{timeout}{Time in seconds to keep the wallet unlocked.}
}
\description{
Unlock the wallet with a passphrase
}
