% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\name{mc_unsubscribe}
\alias{mc_unsubscribe}
\title{Unsubscribe from assets or streams}
\usage{
mc_unsubscribe(conn, entities, purge = FALSE)
}
\arguments{
\item{conn}{Connection object.}

\item{entities}{A single name/ref/txid or a vector of them.}

\item{purge}{If TRUE (default FALSE), any off-chain data retrieved for this stream will be purged.}
}
\value{
NULL if successful.
}
\description{
Instructs the node to stop tracking one or more asset(s) or stream(s).
}
