% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{mc_update}
\alias{mc_update}
\title{Update asset status (open/closed)}
\usage{
mc_update(conn, asset, params)
}
\arguments{
\item{conn}{Connection object.}

\item{asset}{Asset name, ref or txid.}

\item{params}{List, e.g., list(open = FALSE).}
}
\description{
Update asset status (open/closed)
}
