% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{mc_update_from}
\alias{mc_update_from}
\title{Update asset status from specific address}
\usage{
mc_update_from(conn, from_address, asset, params)
}
\arguments{
\item{conn}{A connection object created by \code{\link{mc_connect}}.}

\item{from_address}{Character string. Address that pays for the transaction.}

\item{asset}{Character string. Asset name, reference, or issuance ID.}

\item{params}{A list of parameters to update.}
}
\value{
A character string containing the transaction ID.
}
\description{
Changes the status of an asset (e.g., open/close) and specifies the address
that pays for and authorizes the update.
}
\examples{
\dontrun{
# Close the asset from a specific address
txid <- mc_update_from(conn, "1A...", "mycoin", list(open = FALSE))
}

}
\seealso{
\code{\link{mc_update}} for simpler usage.

Other assets: 
\code{\link{mc_get_asset_info}()},
\code{\link{mc_get_token_info}()},
\code{\link{mc_issue}()},
\code{\link{mc_issue_from}()},
\code{\link{mc_issue_more}()},
\code{\link{mc_issue_more_from}()},
\code{\link{mc_issue_token}()},
\code{\link{mc_issue_token_from}()},
\code{\link{mc_list_asset_issues}()},
\code{\link{mc_list_assets}()},
\code{\link{mc_update}()}
}
\concept{assets}
