% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null_default}
\alias{null_default}
\alias{\%||\%}
\title{Null-default operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{An object to check for \code{NULL}.}

\item{b}{The default value to return if \code{a} is \code{NULL}.}
}
\value{
\code{a} if it is not \code{NULL}, otherwise \code{b}.
}
\description{
This infix operator provides a convenient way to handle \code{NULL} values
by providing a default value.
}
\keyword{internal}
