% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null_default}
\alias{null_default}
\alias{\%||\%}
\title{Null-default operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{An object}

\item{b}{An object}
}
\value{
a if not null, else b
}
\description{
Null-default operator
}
\keyword{internal}
