% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{print.mc_help}
\alias{print.mc_help}
\title{Print MultiChain help}
\usage{
\method{print}{mc_help}(x, ...)
}
\arguments{
\item{x}{An mc_help object.}

\item{...}{Additional arguments.}
}
\description{
Print MultiChain help
}
