% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rpc_res_to_df}
\alias{rpc_res_to_df}
\title{Convert RPC result list to data frame}
\usage{
rpc_res_to_df(res)
}
\arguments{
\item{res}{A list of objects or primitives returned by \code{\link[=mc_rpc]{mc_rpc()}}.}
}
\value{
A data frame. Returns an empty data frame if the input is empty.
}
\description{
Convert RPC result list to data frame
}
\keyword{internal}
